fix: Template the error page
404ページとその他エラーページのtemplateを作成しまとめる
